More Tile Selection Features and Frontend
Removed the Selectetd Tiles resource in favor of entities. While this does make it harder to know the order tiles were selected in, that's a problem for another day if I ever decide that feature is needed.

I must manually ensure that a tile is not both selectable and selected at the same time.

Selection Critera now allows for me to specify what tiles are valid.

Also, I added floating UI bits to visually indicate which tiles are selectable.